separate out test cases into individual tests
